Added a PGN selection window, which allows you to select (the first game from) the PGNs in the application's home directory.
At the moment, the implementation is highly inefficient on the actual device for large PGN files, since it parses the whole thing at once. Work needs to be done to only parse "stubs" in the first pass, and only parse the moves themselves once the user has selected the game.